refactor(factory): extract assembly validators to shared lib module
scripts/__tests__/unit/assemble-validation.test.js had a local copy
of SAFE_PLACEHOLDER_PATTERNS (10 entries) that had drifted behind
the real one in scripts/assemble-factory.js (16+ entries after
Tasks 3-5 additions). The test's header comment said it "mirrors
the logic in assemble.js and assemble-factory.js" — true at the
time it was written, no longer true.

Extract SAFE_PLACEHOLDER_PATTERNS + validateFactoryTemplate +
validateFactoryAssembly into scripts/lib/factory-assembly-validation.js
and import from both:
  - scripts/assemble-factory.js (the CLI)
  - scripts/__tests__/unit/assemble-validation.test.js (the unit test)

Now they share one source of truth. No functional change; tests
still 647 pass / 36 fail. Smoke test via sell-ready fixture still
produces a 218 KB index.html with 0 use-fireproof, 3 subscribe-cta,
0 claim-bar.
